feat(track): Only track when prod
